Use store defaults for Explore atom
